requirements.txt
Flask==0.12.2
azure-storage==0.36.0
keras==2.1.3
tensorflow==1.4.0
pandas==0.20.3
numpy==1.13.3
h5py
